requirements.txt
# AI Generation Dependencies
httpx>=0.25.0
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0

# Optional: for async file operations
anyio>=4.0.0
